requirements.txt
fastapi
uvicorn
python-dotenv
requests
firebase-admin
google-cloud-firestore